Remove defaults, remove debugger
